+++ pytsite/core/assetman/__init__.py
@@ -1,4 +1,4 @@
-"""Assetman Plugin Init
+"""Assetman Plugin Init.
 """
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -24,4 +24,4 @@ __init()
 
 
 # Public API
-from ._functions import register_package, add, add_js, add_css, dump_js, dump_css
+from ._functions import register_package, add, add_location, remove, remove_location, dump_js, dump_css