+++ pytsite/core/assetman/_functions.py
@@ -4,24 +4,25 @@ __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-from os import path
-from importlib.util import find_spec
-from pytsite.core import router
+from os import path as _path
+from importlib.util import find_spec as _find_spec
+from pytsite.core import router as _router
 
 __packages = {}
 __global_links = {'css': [], 'js': []}
 __pattern_links = {}
-__links = {}
+__exact_links = {}
+
 
 def register_package(package_name: str, assets_dir: str='assets'):
     """Register assets container.
     """
-    spec = find_spec(package_name)
+    spec = _find_spec(package_name)
     if not spec:
         raise Exception("Package '{}' is not found.".format(package_name))
 
-    dir_path = path.join(path.dirname(spec.origin), assets_dir)
-    if not path.isdir(dir_path):
+    dir_path = _path.join(_path.dirname(spec.origin), assets_dir)
+    if not _path.isdir(dir_path):
         FileNotFoundError("Directory '{}' is not found.".format(dir_path))
 
     __packages[package_name] = dir_path
@@ -32,17 +33,20 @@ def get_packages() -> dict:
 
 
 def add_location(location: str, collection: str, route_path: str=None):
-    """Add an asset location to the collection.
+    """Add an asset location.
     """
     if not location.startswith('http'):
         if not location.startswith('app') and not location.startswith('pytsite.'):
             location = 'pytsite.' + location
 
     if not route_path:
-        route_path = router.current_path(True)
+        route_path = _router.current_path(True)
 
+    # Any path
     if route_path == '*':
         __global_links[collection].append(location)
+
+    # Prefixed paths group
     elif route_path.endswith('*'):
         if route_path.endswith('/*'):
             route_path = route_path.replace('/*', '*')
@@ -50,39 +54,68 @@ def add_location(location: str, collection: str, route_path: str=None):
             __pattern_links[route_path] = {'css': [], 'js': []}
         if location not in __pattern_links[route_path][collection]:
             __pattern_links[route_path][collection].append(location)
+
+    # Exact path
     else:
-        if route_path not in __links:
-            __links[route_path] = {'css': [], 'js': []}
-        if location not in __links[route_path][collection]:
-            __links[route_path][collection].append(location)
+        if route_path not in __exact_links:
+            __exact_links[route_path] = {'css': [], 'js': []}
+        if location not in __exact_links[route_path][collection]:
+            __exact_links[route_path][collection].append(location)
 
 
-def add_js(location: str, route_path: str=None):
-    """Add a JS asset location to the collection.
+def remove_location(location: str, collection: str, route_path: str=None):
+    """Remove an asset location.
     """
-    add_location(location, 'js', route_path)
+    if not location.startswith('http'):
+        if not location.startswith('app') and not location.startswith('pytsite.'):
+            location = 'pytsite.' + location
 
+    if not route_path:
+        route_path = _router.current_path(True)
 
-def add_css(location: str, route_path: str=None):
-    """Add a CSS asset location to the collection.
-    """
-    add_location(location, 'css', route_path)
+    # Any path
+    if route_path == '*':
+        __global_links[collection] = [em for em in __global_links[collection] if em != location]
+
+    # Prefixed paths group
+    elif route_path.endswith('*'):
+        if route_path.endswith('/*'):
+            route_path = route_path.replace('/*', '*')
+        if route_path not in __pattern_links:
+            return
+        __pattern_links[route_path][collection] = [em for em in __pattern_links[route_path][collection]
+                                                   if em != location]
+
+    # Exact path
+    else:
+        if route_path not in __exact_links:
+            return
+        __exact_links[route_path][collection] = [em for em in __exact_links[route_path][collection] if em != location]
 
 
 def add(location: str, route_path: str=None):
     """Shortcut.
     """
     if location.endswith('.js'):
-        add_js(location, route_path)
+        add_location(location, 'js', route_path)
     elif location.endswith('.css'):
-        add_css(location, route_path)
+        add_location(location, 'css', route_path)
+    else:
+        raise ValueError("Cannot detect collection to add for '{}'.".format(location))
+
+
+def remove(location: str, route_path: str=None):
+    if location.endswith('.js'):
+        remove_location(location, 'js', route_path)
+    elif location.endswith('.css'):
+        remove_location(location, 'css', route_path)
     else:
         raise ValueError("Cannot detect collection to add for '{}'.".format(location))
 
 
 def get_locations(collection: str) -> list:
     r = []
-    current_path = router.current_path(True)
+    current_path = _router.current_path(True)
 
     # Links for every path
     for location in __global_links[collection]:
@@ -95,13 +128,14 @@ def get_locations(collection: str) -> list:
                 if location not in r:
                     r.append(location)
 
-    # Links for exact path
-    if current_path in __links:
-        for location in __links[current_path][collection]:
+    # Links for exact paths
+    if current_path in __exact_links:
+        for location in __exact_links[current_path][collection]:
             if location not in r:
                 r.append(location)
 
     return r
+
 
 def dump_js() -> str:
     """Dump JS links.
@@ -129,7 +163,7 @@ def get_url(location: str) -> str:
     if location.startswith('http'):
         return location
     package_name, asset_path = __split_asset_location_info(location)
-    return router.url('/assets/{}/{}'.format(package_name, asset_path), strip_lang=True)
+    return _router.url('/assets/{}/{}'.format(package_name, asset_path), strip_lang=True)
 
 
 def __split_asset_location_info(location: str) -> dict: