+++ pytsite/core/validation/_error.py
@@ -6,4 +6,9 @@ __license__ = 'MIT'
 
 
 class ValidationError(ValueError):
-    pass
+    def __init__(self, msg_args: dict=None):
+        self._msg_args = msg_args if msg_args else {}
+
+    @property
+    def msg_args(self) -> dict:
+        return self._msg_args