+++ pytsite/core/widget/assets/js/datetime.js
@@ -1,5 +1,5 @@
 $(function() {
-    $('.widget-datetime-input').each(function() {
+    $('.widget-select-datetime').each(function() {
         var widget = $(this);
         widget.find('input').datetimepicker({
             lang: pytsite.lang.current,