+++ pytsite/geo/_widget.py
@@ -17,11 +17,14 @@ class SearchAddress(_widget.Base):
         """
         super().__init__(**kwargs)
 
+        self._autodetect = kwargs.get('autodetect', False)
+
         lng = _lang.get_current_lang()
-        _assetman.add_js('https://maps.googleapis.com/maps/api/js?libraries=places&language=' + lng)
+        _assetman.add_location('https://maps.googleapis.com/maps/api/js?libraries=places&language=' + lng, 'js')
         _assetman.add('geo@js/address-input.js')
 
         self._group_cls += ' widget-geo-address-input'
+        self._group_data['autodetect'] = int(self._autodetect)
 
     def set_value(self, val: dict, **kwargs: dict):
         """Set value of the widget.